+++ Common/changelog.md
@@ -4,6 +4,7 @@
   - Foo
 - Tweaked display settings of player plushies
 - Made player plushies display last in creative tab
+- Added 4 new player plushies
 
 ## Version 1.2.1
 - Fixed rare player plushie not visible